move main to cmd directory
