+++ .goreleaser.yaml
@@ -10,6 +10,7 @@ builds:
       - linux
       - windows
       - netbsd
+    main: ./cmd/main.go
 archives:
   - name_template: "{{ .Os }}-{{ .Arch }}"
     format: binary
