--- cmd/workflow.go
@@ -1,10 +1,0 @@
-package cmd
-
-import (
-	"github.com/mdb/gh-dispatch/dispatch"
-)
-
-func init() {
-	workflowCmd := dispatch.NewCmdWorkflow()
-	rootCmd.AddCommand(workflowCmd)
-}
